# Escora-Admin-Portal-main/artifacts/api-server/src/routes/orders.ts
import { Router, type IRouter } from "express";
import { eq, and, type SQL } from "drizzle-orm";
import { db, ordersTable } from "@workspace/db";
import {
  ListOrdersQueryParams,
  ListOrdersResponse,
  CreateOrderBody,
  GetOrderParams,
  GetOrderResponse,
  UpdateOrderParams,
  UpdateOrderBody,
  UpdateOrderResponse,
  DeleteOrderParams,
} from "@workspace/api-zod";
import { logger } from "../lib/logger";
import { sendBookingConfirmation, sendAdminBookingNotification } from "../lib/email";

const router: IRouter = Router();

function serializeDates<T extends { createdAt: Date; updatedAt: Date }>(row: T) {
  return { ...row, createdAt: row.createdAt.toISOString(), updatedAt: row.updatedAt.toISOString() };
}

async function lookupGeo(ip: string): Promise<{ geoCountry?: string; geoRegion?: string; geoCity?: string }> {
  // Skip loopback / private addresses
  if (!ip || ip === "::1" || ip.startsWith("127.") || ip.startsWith("192.168.") || ip.startsWith("10.") || ip === "::ffff:127.0.0.1") {
    return {};
  }
  try {
    const cleanIp = ip.replace(/^::ffff:/, "");
    const res = await fetch(`http://ip-api.com/json/${cleanIp}?fields=status,country,regionName,city`, {
      signal: AbortSignal.timeout(3000),
    });
    if (!res.ok) return {};
    const data = await res.json() as { status: string; country?: string; regionName?: string; city?: string };
    if (data.status !== "success") return {};
    return { geoCountry: data.country, geoRegion: data.regionName, geoCity: data.city };
  } catch {
    return {};
  }
}

router.get("/orders/export", async (req, res): Promise<void> => {
  const query = ListOrdersQueryParams.safeParse(req.query);
  if (!query.success) {
    res.status(400).json({ error: query.error.message });
    return;
  }

  const conditions: SQL[] = [];
  if (query.data.status !== undefined) {
    conditions.push(eq(ordersTable.status, query.data.status));
  }

  const orders = conditions.length
    ? await db.select().from(ordersTable).where(and(...conditions)).orderBy(ordersTable.createdAt)
    : await db.select().from(ordersTable).orderBy(ordersTable.createdAt);

  const headers = [
    "ID","Guest Name","Email","Phone","Travel Style","Destinations","Travel Date","Departure Date",
    "Guests","Budget Range","Status","Total Amount (INR)","Package","Special Requests",
    "UTM Source","UTM Medium","UTM Campaign","Country","City","Created At"
  ];

  function escapeCsv(v: string | number | null | undefined): string {
    if (v === null || v === undefined) return "";
    const s = String(v);
    if (s.includes(",") || s.includes('"') || s.includes("\n")) {
      return `"${s.replace(/"/g, '""')}"`;
    }
    return s;
  }

  const rows = orders.map(o => [
    o.id, o.guestName, o.guestEmail, o.guestPhone, o.travelStyle, o.destinations,
    o.travelDate, o.departureDate, o.guestCount, o.budgetRange, o.status, o.totalAmount,
    o.packageName, o.specialRequests, o.utmSource, o.utmMedium, o.utmCampaign,
    o.geoCountry, o.geoCity, o.createdAt.toISOString(),
  ].map(escapeCsv).join(","));

  const csv = [headers.join(","), ...rows].join("\r\n");

  res.setHeader("Content-Type", "text/csv");
  res.setHeader("Content-Disposition", `attachment; filename="enquiries-${new Date().toISOString().slice(0,10)}.csv"`);
  res.send(csv);
});

router.get("/orders", async (req, res): Promise<void> => {
  const query = ListOrdersQueryParams.safeParse(req.query);
  if (!query.success) {
    res.status(400).json({ error: query.error.message });
    return;
  }

  const conditions: SQL[] = [];
  if (query.data.status !== undefined) {
    conditions.push(eq(ordersTable.status, query.data.status));
  }

  try {
    const orders = conditions.length
      ? await db.select().from(ordersTable).where(and(...conditions)).orderBy(ordersTable.createdAt)
      : await db.select().from(ordersTable).orderBy(ordersTable.createdAt);
    res.json(ListOrdersResponse.parse(orders.map(serializeDates)));
  } catch {
    res.json([]);
  }
});

router.post("/orders", async (req, res): Promise<void> => {
  const parsed = CreateOrderBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const ipAddress = (req.ip ?? "").replace(/^::ffff:/, "") || undefined;
  const userAgent = req.headers["user-agent"] || undefined;

  // Fire-and-forget geo lookup so it doesn't slow the response
  const geoPromise = ipAddress ? lookupGeo(ipAddress) : Promise.resolve({});

  let geo: { geoCountry?: string; geoRegion?: string; geoCity?: string } = {};
  try {
    geo = await geoPromise;
  } catch {
    logger.warn("Geo lookup failed");
  }

  const [order] = await db.insert(ordersTable).values({
    ...parsed.data,
    ipAddress,
    userAgent,
    ...geo,
  }).returning();

  // Send emails asynchronously — don't block the response
  Promise.all([
    sendBookingConfirmation(order),
    sendAdminBookingNotification(order),
  ]).catch(() => { /* non-fatal */ });

  res.status(201).json(GetOrderResponse.parse(serializeDates(order)));
});

router.get("/orders/:id", async (req, res): Promise<void> => {
  const params = GetOrderParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [order] = await db.select().from(ordersTable).where(eq(ordersTable.id, params.data.id));
  if (!order) {
    res.status(404).json({ error: "Order not found" });
    return;
  }
  res.json(GetOrderResponse.parse(serializeDates(order)));
});

router.patch("/orders/:id", async (req, res): Promise<void> => {
  const params = UpdateOrderParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const parsed = UpdateOrderBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const [order] = await db
    .update(ordersTable)
    .set({ ...parsed.data, updatedAt: new Date() })
    .where(eq(ordersTable.id, params.data.id))
    .returning();
  if (!order) {
    res.status(404).json({ error: "Order not found" });
    return;
  }
  res.json(UpdateOrderResponse.parse(serializeDates(order)));
});

router.delete("/orders/:id", async (req, res): Promise<void> => {
  const params = DeleteOrderParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [order] = await db.delete(ordersTable).where(eq(ordersTable.id, params.data.id)).returning();
  if (!order) {
    res.status(404).json({ error: "Order not found" });
    return;
  }
  res.sendStatus(204);
});

export default router;