# Escora-Admin-Portal-main/lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /packages:
    get:
      operationId: listPackages
      summary: List all packages
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: published
          schema:
            type: boolean
        - in: query
          name: featured
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      summary: Create a package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    get:
      operationId: getPackage
      summary: Get package by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "404":
          description: Not found
    patch:
      operationId: updatePackage
      summary: Update a package
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      summary: Delete a package
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Itineraries ───────────────────────────────────────────────
...
  /travel-categories:
    get:
      operationId: listTravelCategories
      summary: List all travel categories
      parameters:
        - in: query
          name: published
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TravelCategory"
    post:
      operationId: createTravelCategory
      summary: Create a travel category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelCategory"
  /travel-categories/{id}:
    get:
      operationId: getTravelCategory
      summary: Get travel category by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelCategory"
        "404":
          description: Not found
    patch:
      operationId: updateTravelCategory
      summary: Update a travel category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelCategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelCategory"
    delete:
      operationId: deleteTravelCategory
      summary: Delete a travel category
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Orders / Enquiries ────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      summary: List all orders/enquiries
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      summary: Submit an enquiry or booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      summary: Get order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      summary: Update order status or details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      summary: Delete an order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Contact Enquiries ────────────────────────────────────────
...
components:
  schemas:
...
    TravelCategory:
      type: object
      required: [id, slug, label, sortOrder, published, createdAt, updatedAt]
      properties:
        id:
          type: integer
        slug:
          type: string
        label:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        priceFrom:
          type: ["string", "null"]
        priceLabel:
          type: ["string", "null"]
        journeyCount:
          type: ["string", "null"]
        sortOrder:
          type: integer
        published:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    TravelCategoryInput:
      type: object
      required: [slug, label]
      properties:
        slug:
          type: string
        label:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        priceFrom:
          type: string
        priceLabel:
          type: string
        journeyCount:
          type: string
        sortOrder:
          type: integer
        published:
          type: boolean
    TravelCategoryUpdate:
      type: object
      properties:
        slug:
          type: string
        label:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        priceFrom:
          type: string
        priceLabel:
          type: string
        journeyCount:
          type: string
        sortOrder:
          type: integer
        published:
          type: boolean

    # ── Package ──────────────────────────────────────────────────
    Package:
      type: object
      required: [id, name, slug, category, durationNights, published, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        shortDesc:
          type: ["string", "null"]
        durationNights:
          type: integer
        route:
          type: ["string", "null"]
        bestFor:
          type: ["string", "null"]
        season:
          type: ["string", "null"]
        priceFrom:
          type: ["integer", "null"]
        heroImageUrl:
          type: ["string", "null"]
        galleryImages:
          type: ["string", "null"]
        itinerary:
          type: ["string", "null"]
        inclusions:
          type: ["string", "null"]
        exclusions:
          type: ["string", "null"]
        faqs:
          type: ["string", "null"]
        accommodations:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        featured:
          type: boolean
        published:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    PackageInput:
      type: object
      required: [name, slug, category, durationNights]
      properties:
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        description:
          type: string
        shortDesc:
          type: string
        durationNights:
          type: integer
        route:
          type: string
        bestFor:
          type: string
        season:
          type: string
        priceFrom:
          type: integer
        heroImageUrl:
          type: string
        galleryImages:
          type: string
        itinerary:
          type: string
        inclusions:
          type: string
        exclusions:
          type: string
        faqs:
          type: string
        accommodations:
          type: string
        tags:
          type: string
        featured:
          type: boolean
        published:
          type: boolean
    PackageUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        description:
          type: string
        shortDesc:
          type: string
        durationNights:
          type: integer
        route:
          type: string
        bestFor:
          type: string
        season:
          type: string
        priceFrom:
          type: integer
        heroImageUrl:
          type: string
        galleryImages:
          type: string
        itinerary:
          type: string
        inclusions:
          type: string
        exclusions:
          type: string
        faqs:
          type: string
        accommodations:
          type: string
        tags:
          type: string
        featured:
          type: boolean
        published:
          type: boolean

    # ── Itinerary ────────────────────────────────────────────────
...
    Order:
      type: object
      required: [id, guestName, guestEmail, guestPhone, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        packageId:
          type: ["integer", "null"]
        packageName:
          type: ["string", "null"]
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        travelStyle:
          type: ["string", "null"]
        destinations:
          type: ["string", "null"]
        travelDate:
          type: ["string", "null"]
        departureDate:
          type: ["string", "null"]
        guestCount:
          type: ["integer", "null"]
        budgetRange:
          type: ["string", "null"]
        specialRequests:
          type: ["string", "null"]
        status:
          type: string
        totalAmount:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        userAgent:
          type: ["string", "null"]
        referrerUrl:
          type: ["string", "null"]
        landingPage:
          type: ["string", "null"]
        utmSource:
          type: ["string", "null"]
        utmMedium:
          type: ["string", "null"]
        utmCampaign:
          type: ["string", "null"]
        browserLanguage:
          type: ["string", "null"]
        screenResolution:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        geoCountry:
          type: ["string", "null"]
        geoRegion:
          type: ["string", "null"]
        geoCity:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderInput:
      type: object
      required: [guestName, guestEmail, guestPhone]
      properties:
        packageId:
          type: integer
        packageName:
          type: string
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        travelStyle:
          type: string
        destinations:
          type: string
        travelDate:
          type: string
        departureDate:
          type: string
        guestCount:
          type: integer
        budgetRange:
          type: string
        specialRequests:
          type: string
        ipAddress:
          type: string
        userAgent:
          type: string
        referrerUrl:
          type: string
        landingPage:
          type: string
        utmSource:
          type: string
        utmMedium:
          type: string
        utmCampaign:
          type: string
        browserLanguage:
          type: string
        screenResolution:
          type: string
        timezone:
          type: string
        geoCountry:
          type: string
        geoRegion:
          type: string
        geoCity:
          type: string
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
        totalAmount:
          type: integer
        notes:
          type: string
        travelStyle:
          type: string
        destinations:
          type: string
        travelDate:
          type: string
        departureDate:
          type: string
        guestCount:
          type: integer
        budgetRange:
          type: string
        specialRequests:
          type: string

    # ── Enquiry ──────────────────────────────────────────────────
...